Lesson 13 - create folder with components(Posts)
