+++ src/App.css
@@ -8,26 +8,3 @@
     grid-gap: 10px;
     width: 1200px;
 }
-
-.header img {
-    width: 60px;
-}
-
-.header {
-    grid-area: header;
-    background-color: green;
-}
-
-.nav {
-    grid-area: nav;
-    background-color: darkgreen;
-}
-
-.nav a {
-    text-decoration: none;
-}
-
-.content {
-    grid-area: content;
-    background-color: aqua;
-}
