+++ src/App.js
@@ -1,8 +1,8 @@
 import './App.css';
 import React from 'react';
-import Header from "./components/Header";
-import Nav from "./components/Nav";
-import Profile from "./components/Profile";
+import Header from "./components/Header/Header";
+import Nav from "./components/Navbar/Nav";
+import Profile from "./components/Profile/Profile";
 
 const App = () => {
     return (
