+++ src/components/Header/Header.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import '../App.css';
+import s from './Header.module.css';
 
 const Header = () => {
     return (
-        <header className='header'>
+        <header className={s.header}>
             <img src='https://static.rfstat.com/renderforest/images/v2/logo-homepage/gradient_4.png' alt=''/>
         </header>
     )
