+++ src/components/Header/Header.module.css
@@ -1,0 +1,8 @@
+.header img {
+    width: 60px;
+}
+
+.header {
+    grid-area: header;
+    background-color: green;
+}
