--- src/components/Nav.jsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import '../App.css'
-
-const Nav = () => {
-    return (
-        <nav className='nav'>
-            <div>
-                <a href='#'>Profile</a>
-            </div>
-            <div>
-                <a href='#'>Messages</a>
-            </div>
-            <div>
-                <a href='#'>Music</a>
-            </div>
-            <div>
-                <a href='#'>News</a>
-            </div>
-        </nav>
-    )
-}
-
-export default Nav;
