+++ src/components/Navbar/Nav.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import s from './Nav.module.css';
+
+const Nav = () => {
+    return (
+        <nav className={s.nav}>
+            <div className={s.item}>
+                <a>Profile</a>
+            </div>
+            <div className={s.item}>
+                <a>Messages</a>
+            </div>
+            <div className={s.item}>
+                <a>Music</a>
+            </div>
+            <div className={s.item}>
+                <a>News</a>
+            </div>
+        </nav>
+    )
+}
+
+export default Nav;
