+++ src/components/Profile/Profile.module.css
@@ -1,0 +1,12 @@
+.content {
+    grid-area: content;
+    background-color: aqua;
+}
+
+img {
+    width: 100%;
+}
+
+.item {
+    color: yellowgreen;
+}
